Refactored plugin and plugin helpers modules
