+++ flint-plugins/report/json/run.lua
@@ -1,3 +1,5 @@
+local json = require("json")
+
 function Run(options)
     local config = options.config
     local output = options.output
