+++ flint/src/plugin/find.rs
@@ -46,7 +46,7 @@ pub fn dir() -> PathBuf {
 
 pub fn list<'a>() -> AppResult<&'a BTreeSet<Plugin>> {
     let lua = Lua::new();
-    add_helper_globals(&lua);
+    add_helper_globals(&lua)?;
 
     if PLUGINS.get().is_some() {
         return Ok(PLUGINS.get().unwrap());
